Fixed persistent debug markers and did some clean-up
