+++ bepinex_dev/SPTQuestingBots/Models/BotPathData.cs
@@ -12,7 +12,7 @@ using UnityEngine;
 
 namespace SPTQuestingBots.Models
 {
-    public enum BotPathUpdateReason
+    public enum BotPathUpdateNeededReason
     {
         None = 0,
         Force,
@@ -35,61 +35,58 @@ namespace SPTQuestingBots.Models
             bot = botOwner;
         }
 
-        public BotPathUpdateReason Update(Vector3 target, float reachDistance = 0.5f, bool force = false)
+        public BotPathUpdateNeededReason CheckIfUpdateIsNeeded(Vector3 target, float reachDistance = 0.5f, bool force = false)
         {
             bool requiresUpdate = false;
-            BotPathUpdateReason reason = BotPathUpdateReason.None;
+            BotPathUpdateNeededReason reason = BotPathUpdateNeededReason.None;
 
             float distanceFromStartPosition = Vector3.Distance(bot.Position, StartPosition);
 
             if (force)
             {
                 requiresUpdate = true;
-                reason = BotPathUpdateReason.Force;
+                reason = BotPathUpdateNeededReason.Force;
             }
 
+            // Check if a new target position has been set or if the reach distance has been modified
             if (!requiresUpdate && ((target != TargetPosition) || (reachDistance != ReachDistance)))
             {
                 TargetPosition = target;
                 ReachDistance = reachDistance;
 
                 requiresUpdate = true;
-                reason = BotPathUpdateReason.NewTarget;
+                reason = BotPathUpdateNeededReason.NewTarget;
             }
 
+            // If the path is incomplete, it should be regularly updated in case Unity is able to resolve the it
             if (!requiresUpdate)
             {
                 if (Status == UnityEngine.AI.NavMeshPathStatus.PathInvalid)
                 {
                     requiresUpdate = true;
-                    reason = BotPathUpdateReason.IncompletePath;
+                    reason = BotPathUpdateNeededReason.IncompletePath;
                 }
                 else if ((Status == UnityEngine.AI.NavMeshPathStatus.PathPartial) && (Time.time - LastSetTime > IncompletePathRefreshDelay))
                 {
                     requiresUpdate = true;
-                    reason = BotPathUpdateReason.IncompletePath;
+                    reason = BotPathUpdateNeededReason.IncompletePath;
                 }
             }
 
-            /*if (!requiresUpdate && (bot.Mover.LastPathSetTime != LastSetTime))
-            {
-                requiresUpdate &= distanceFromStartPosition > MinRefreshDistance;
-                reason = BotPathUpdateReason.RefreshNeededTime;
-            }*/
-
+            // Check if the bot's path has been changed by another component (i.e. Looting Bots, SAIN, etc.)
             if (!requiresUpdate)
             {
                 Vector3[] currentPath = bot.Mover.GetCurrentPath();
                 if (currentPath == null)
                 {
                     requiresUpdate = true;
-                    reason = BotPathUpdateReason.RefreshNeededPath;
+                    reason = BotPathUpdateNeededReason.RefreshNeededPath;
                 }
 
                 if (!requiresUpdate && Corners.Any() && (currentPath?.Any() == true) && (currentPath.Last() != Corners.Last()))
                 {
                     requiresUpdate &= distanceFromStartPosition > MinRefreshDistance;
-                    reason = BotPathUpdateReason.RefreshNeededPath;
+                    reason = BotPathUpdateNeededReason.RefreshNeededPath;
                 }
             }
 
@@ -118,6 +115,7 @@ namespace SPTQuestingBots.Models
             Status = CreatePathSegment(bot.Position, target, out Vector3[] corners);
             if (Status == UnityEngine.AI.NavMeshPathStatus.PathPartial)
             {
+                // Check if any static paths exist to the target position and sort them based on the approximate total path length for the bot
                 BotQuestBuilder botQuestBuilder = Singleton<GameWorld>.Instance.GetComponent<BotQuestBuilder>();
                 IEnumerable<StaticPathData> staticPaths = botQuestBuilder
                     .GetStaticPaths(target)
@@ -130,12 +128,14 @@ namespace SPTQuestingBots.Models
 
                 foreach (StaticPathData staticPath in staticPaths)
                 {
+                    // Check if Unity can form a complete path from the bot to the static path's endpoint
                     UnityEngine.AI.NavMeshPathStatus staticPathStatus = CreatePathSegment(bot.Position, staticPath.StartPosition, out Vector3[] staticPathCorners);
                     if (staticPathStatus == UnityEngine.AI.NavMeshPathStatus.PathComplete)
                     {
                         Corners = staticPathCorners;
                         Status = UnityEngine.AI.NavMeshPathStatus.PathComplete;
 
+                        // Merge the paths and instruct the bot to use the combination
                         StaticPathData combinedPath = Append(staticPath);
                         SetCorners(combinedPath.Corners);
